move adhoc scheduler into it's own file
